# RocksDB increasing background threads to speedup compaction.