RocksDB increasing background threads to speedup compaction.
max_background_compactions = 4;  // Maximum number of concurrent background jobs(threads used for flush and compaction)
max_subcompactions = 2;   // Maximum number of threads that will concurrently perform a compaction job by breaking it into multiple smaller ones that are run simultaneously

Why?
1)To give more cpu resources/time for compaction, as compaction is giving low priority and always postponed to do until we reach the thresholds.
2)Observed that a SS having high(>80 billion bytes) pendingCompactionBytes is not using 100% CPU. So, with this we can use the unused CPU for compaction.
3)We see write stalls upto 50 seconds in the case where we have lot of pending compacted data in L0 of LSM. The data in L0 to L1 mostly cannot be compacted parallely, so setting max_subcompactions to 2 would help.

Testing:
-Splunk Stat graphs posted in fab-storage-layer group.
-Correctness test passed.
-Perf Test Output:
LOAD to fill 30% of cluster:
INSERT_AFTER 106628.902512 ops/sec 50th 1 90th 2 99th 11111 99.9th 1505279 99.99th 2854911 (us) 2nd run
INSERT_BEFORE 115697.527355 ops/sec 50th 1 90th 3 99th 4843 99.9th 957951 99.99th 5222399 (us)
Workload CK:
INSERT_AFTER 36163.807261 ops/sec 50th 2 90th 26 99th 147 99.9th 286 99.99th 585 (us)
INSERT_BEFORE 39019.169467 ops/sec 50th 2 90th 26 99th 142 99.9th 271 99.99th 546 (us)
SCAN_AFTER 501808.156087 ops/sec 50th 1217 90th 2105 99th 8415 99.9th 18831 99.99th 41119 (us)
SCAN_BEFORE 541901.519770 ops/sec 50th 1131 90th 2005 99th 7479 99.9th 14407 99.99th 29919 (us)
Throughtout decreased for both insert and scan. This is expected, as we are using comparatively more CPU for compaction and I think this would beneficial if the load continues to run for long hours.

Splunk Observations:
1)Avg(PendingCompactionBytes) is low
2)Better CPU usage of SS
3)Less write stalls
4)Number of parallel running compactions is 3 in atleast 60% of the workload time.



# Code-Reviewer Section

The general guidelines can be found [here](https://github.com/apple/foundationdb/wiki/FoundationDB-Commit-Process).

Please check each of the following things and check *all* boxes before accepting a PR.

- [ ] The PR has a description, explaining both the problem and the solution.
- [ ] The description mentions which forms of testing were done and the testing seems reasonable.
- [ ] Every function/class/actor that was touched is reasonably well documented.

## For Release-Branches

If this PR is made against a release-branch, please also check the following:

- [ ] This change/bugfix is a cherry-pick from the next younger branch (younger `release-branch` or `master` if this is the youngest branch)
- [ ] There is a good reason why this PR needs to go into a release branch and this reason is documented (either in the description above or in a linked GitHub issue)